Changing class name and adding u+alpha designs
